fix some errors

+++ database/migrations/2019_09_15_000020_create_domains_table.php
@@ -13,10 +13,14 @@ class CreateDomainsTable extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('domains')) {
+            return;
+        }
+
         Schema::create('domains', function (Blueprint $table) {
             $table->increments('id');
             $table->string('domain', 255)->unique();
-            $table->string('tenant_id');
+            $table->unsignedBigInteger('tenant_id');
 
             $table->timestamps();
             $table->foreign('tenant_id')->references('id')->on('tenants')->onUpdate('cascade')->onDelete('cascade');
